feat: Implementando permissões de acesso no sistema de criação de corrida. Melhorando página de visualização de corridas

+++ app/Http/Controllers/RaceController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Race; // Importando o modelo Race
+use Illuminate\Support\Facades\Auth; // Importando o facade Auth
 use Illuminate\Http\Request;
 
 class RaceController extends Controller
@@ -17,16 +18,24 @@ class RaceController extends Controller
     // Exibir o formulário para criar uma nova corrida
     public function create()
     {
+        $user = Auth::user();
+        if(!$user->is_administrator){
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página');
+        }
         return view('races.create'); // Retorna a view para criar uma nova corrida
     }
 
     // Armazenar uma nova corrida no banco de dados
     public function store(Request $request)
     {
+        $user = Auth::user();
+        if(!$user->is_administrator){
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
-            'category' => 'required|string|max:255',
-            'max_vehicles' => 'required|integer',
+            'category' => 'required|string|max:1',
+            'max_vehicles' => 'required|integer|max:10',
             'date' => 'required|date',
             'start_location' => 'required|string|max:255',
             'start_latitude' => 'required|numeric',   // Validação da latitude de largada
@@ -50,12 +59,20 @@ class RaceController extends Controller
     // Exibir o formulário para editar uma corrida existente
     public function edit(Race $race)
     {
+        $user = Auth::user();
+        if(!$user->is_administrator){
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página');
+        }
         return view('races.edit', compact('race')); // Passa a corrida para a view de edição
     }
 
     // Atualizar uma corrida existente no banco de dados
     public function update(Request $request, Race $race)
     {
+        $user = Auth::user();
+        if(!$user->is_administrator){
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página');
+        }
         $request->validate([
             'name' => 'required|string|max:255',
             'category' => 'required|string|max:255',
@@ -76,6 +93,10 @@ class RaceController extends Controller
     // Remover uma corrida do banco de dados
     public function destroy(Race $race)
     {
+        $user = Auth::user();
+        if(!$user->is_administrator){
+            return redirect('/')->with('msg', 'Você não tem permissão para acessar essa página');
+        }
         $race->delete(); // Exclui a corrida
         return redirect()->route('races.index')->with('success', 'Corrida excluída com sucesso!'); // Redireciona para a lista de corridas
     }
